+++ src/rome.standalone.js
@@ -1,7 +1,8 @@
 var rome = require('./rome');
+var momentum = require('./momentum');
 
-rome.moment = global.moment;
+rome.use(global.moment);
 
-if (rome.moment === void 0) {
+if (momentum.moment === void 0) {
   throw new Error('rome depends on moment.js, you can get it at http://momentjs.com, or you could use the bundled distribution file instead.');
 }