+++ src/use.js
@@ -1,0 +1,9 @@
+'use strict';
+
+var momentum = require('./momentum');
+
+function use (moment) {
+  momentum.moment = moment;
+}
+
+module.exports = use;